move audio cue config to new file
